+++ host/lib/usrp/usrp2/codec_ctrl.hpp
@@ -54,14 +54,14 @@ public:
      * \param gain from 0-6dB
      */
 
-    virtual void set_rx_digital_gain(float gain) = 0;
+    virtual void set_rx_digital_gain(double gain) = 0;
 
     /*!
      * Set the digital gain correction on the USRP2+ ADC (ADS62P44).
      * \param gain from 0-0.5dB
      */
 
-    virtual void set_rx_digital_fine_gain(float gain) = 0;
+    virtual void set_rx_digital_fine_gain(double gain) = 0;
 
 };
